+++ src/js/GameObjects/gameObject.js
@@ -1,7 +1,6 @@
 import { CONSTANTS } from "../constants.js";
 
 export class GameObject {
-
     constructor(x, y, width, color, vx, vy, takeable){
         this.x = x;
         this.y = y;
@@ -11,7 +10,6 @@ export class GameObject {
         this.vy = vy;
         this.takeable = takeable;
     }
-
     move(dx, dy){
         this.x += dx;
         this.y += dy;
@@ -22,12 +20,11 @@ export class GameObject {
     }
     draw(ctx){
     }
-
 }
+
 export class TakeObject extends GameObject{ //Klasse fuer alle nehmbaren Objekte
     constructor(x,y,width, color, vx,vy){
         super(x, y, width, color, vx, vy, true);
-        this.human = false;
     }
     
     take(){
@@ -37,20 +34,16 @@ export class TakeObject extends GameObject{ //Klasse fuer alle nehmbaren Objekte
             this.x>playerx-this.width
             &&this.y<playery+playerwidth+1
             &&this.y>playery-this.width);
-        
     }
 }
 
-
 export class GroundObject extends GameObject{
-
     constructor(x, y, width, color, vx, vy){
         super(x,y,width, color,vx,vy, false);
         this.x = x;
         this.y = y;
         this.width = width;
         this.color = color;
-        this.human = false;
     }
     draw(ctx){
         ctx.fillStyle = this.color;
@@ -61,143 +54,5 @@ export class GroundObject extends GameObject{
             this.x>playerx-this.width
             &&this.y<playery+playerwidth+1
             &&this.y>playery-this.width);
-        
-    }
-}
-
-
-export class Virus extends TakeObject{
-
-    constructor(x, y, width, outerColor, innerColor,vx,vy){
-        super(x,y,width,innerColor,vx,vy);
-        this.outerColor = outerColor;
-        this.innerColor = innerColor;
-    }
-    draw(ctx){
-        
-        let teil = (Math.PI/10);
-        for(let i = 0; i<20; i+=2){
-            
-            ctx.beginPath();
-            let stueck =teil*i;
-            ctx.arc(this.x, this.y, this.width+2, stueck, stueck+teil);
-            ctx.strokeStyle = this.outerColor;
-            ctx.stroke();
-            ctx.lineWidth =3;
-            ctx.fill();
-        }
-        ctx.shadowBlur = 50;
-        ctx.shadowColor = CONSTANTS.VshadowColor;
-        
-        ctx.beginPath();
-        ctx.arc(this.x, this.y, this.width, 0, 2 * Math.PI);
-        ctx.fillStyle = this.innerColor;
-        ctx.fill();
-        ctx.shadowBlur = 0;
-    }
-    collision(playerx,playery,playerwidth){
-        return (playerx>this.x-this.width-playerwidth&&
-            playerx<this.x+this.width&&
-            playery<this.y+this.width&&
-            playery>this.y-playerwidth-this.width);
-    }
-    take(){
-        let audio = new Audio("../../src/sounds/virus.wav");
-        audio.play();
-        return "1Virus";
-    }
-}
-
-export class Toiletpaper extends TakeObject{
-    constructor(x,y,width, vx,vy){
-        super(x,y, width,CONSTANTS.TPouterColor, vx,vy);
-        this.width = width;
-    }
-    draw(ctx){
-        for(let i= this.width;i>5; i-=4){
-            ctx.beginPath();
-            ctx.arc(this.x, this.y, i, 0, 2 * Math.PI);
-            ctx.fillStyle = CONSTANTS.TPouterColor;
-            ctx.fill();
-            ctx.lineWidth = 0.7;
-            ctx.strokeStyle = CONSTANTS.TPlineColor;
-            ctx.stroke();
-        }
-        ctx.beginPath();
-            ctx.arc(this.x, this.y, 6, 0, 2 * Math.PI);
-            ctx.fillStyle = CONSTANTS.TPinnerColor;
-            ctx.fill();
-        
-    }
-    push(dx, dy){S
-        super.push(dx,dy);
-    }
-    update(){
-        super.update();
-    }
-    take(){
-        
-        let audio = new Audio("../../src/sounds/toiletp.wav");
-        audio.play();
-        return "1TP";
-        
-    }
-    collision(playerx,playery,playerwidth){
-        return (playerx>this.x-this.width-playerwidth&&
-            playerx<this.x+this.width&&
-            playery<this.y+this.width&&
-            playery>this.y-playerwidth-this.width);
-    }
-}
-export class Neighbour extends GameObject{
-    constructor(x,y, radius, facecolor, vx,vy){
-        super(x, y, radius, facecolor, vx,vy, false);
-        this.x = x;
-        this.y = y;
-        this.width = radius;
-        this.facecolor = facecolor;
-        this.human = true;
-        this.met = false;
-    }
-    draw(ctx){
-        ctx.beginPath();
-        ctx.arc(this.x, this.y, this.width, 0, 2 * Math.PI);
-        ctx.fillStyle = this.facecolor;
-        ctx.fill();
-        for(let h = 0; h<6; h++){
-            let hairX = this.width*Math.cos((40+(h*4))/9);
-            let hairY = this.width*Math.sin((40+(h*4))/9);
-            ctx.beginPath();
-            ctx.arc(hairX+this.x, hairY+this.y, 10, 0, 2*Math.PI);
-            ctx.fillStyle = CONSTANTS.NBhairColor;
-            ctx.fill();
-        }
-        ctx.beginPath();
-        ctx.moveTo(this.x-12, this.y-6);
-        ctx.arc(this.x-10, this.y-4, 6, 0, 2 * Math.PI);
-        ctx.fillStyle = "black";
-        ctx.fill();
-        
-        ctx.beginPath();
-        ctx.strokeStyle= "black";
-        ctx.arc(this.x-15, this.y, 15, 0.2, 0.6 * Math.PI);
-        ctx.stroke();
-
-        
-    }
-    meet(){
-        if(!this.met){
-            this.met = true;
-            return "1NB";
-        }
-        //let audio = new Audio("../../src/sounds/altermann.wav");
-        //audio.play();
-        
-    }
-    collision(playerx,playery,playerwidth){
-        return (playerx>this.x-this.width-playerwidth&&
-            playerx<this.x+this.width&&
-            playery<this.y+this.width&&
-            playery>this.y-playerwidth-this.width);
     }
 }
